+++ Tests/INCommonsTests/Swift/FileManager+Directory/FileManager+DirectoryTests.swift
@@ -2,8 +2,23 @@ import INCommons
 import XCTest
 
 class FileManager_DirectoryTests: XCTestCase {
-	func testDocumentDirectoryEndsWithDocuments() {
-		let result = FileManager().documentDirectory
+	func testHomeDirectoryEndsWithSlash() {
+		let result = FileManager.homeDirectory
+		XCTAssertTrue(result.absoluteString.hasSuffix("/"))
+	}
+
+	func testDocumentsDirectoryEndsWithDocuments() {
+		let result = FileManager.documentsDirectory
 		XCTAssertTrue(result.absoluteString.hasSuffix("/Documents/"))
+	}
+
+	func testCachesDirectoryEndsWithLibraryCaches() {
+		let result = FileManager.cachesDirectory
+		XCTAssertTrue(result.absoluteString.hasSuffix("/Library/Caches/"))
+	}
+
+	func testTmpDirectoryEndsWithTmp() {
+		let result = FileManager.tmpDirectory
+		XCTAssertTrue(result.absoluteString.hasSuffix("/tmp/"))
 	}
 }